# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es una API generada automáticamente con Swaggo.",
        "title": "nbox API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Norlis Viamonte",
            "url": "http://www.example.com/support",
            "email": "norlis.viamonte@gmail.com"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/entry": {
            "post": {
                "description": "insert / update vars",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Upsert entries",
                "parameters": [
                    {
                        "description": "Upsert template",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_entry.Entry"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/entry/key": {
            "get": {
                "description": "detail",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Retrieve key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key path",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_entry.Entry"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "description": "delete keys \u0026 children",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key path",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/entry/lookup": {
            "post": {
                "description": "Retrieve values and metadata for a list of keys in a single request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Retrieve multiple entries",
                "parameters": [
                    {
                        "description": "List of keys to retrieve",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/internal_entry.Entry"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/entry/prefix": {
            "get": {
                "description": "list all keys by path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Filter by prefix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key path",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_entry.Entry"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/entry/resolve": {
            "get": {
                "description": "plain value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entry"
                ],
                "summary": "Resolve value",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key path",
                        "name": "v",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_entry.Entry"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BasicAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/me/permissions": {
            "get": {
                "description": "Returns the permission names and resource patterns granted to the authenticated caller's roles (UI hints, not a security boundary).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Current caller capabilities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/problem.Detail"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "internal_entry.Entry": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "development/service/var-example"
                },
                "metadata": {
                    "$ref": "#/definitions/internal_entry.Metadata"
                },
                "secure": {
                    "type": "boolean",
                    "example": false
                },
                "value": {
                    "type": "string",
                    "example": "value 123"
                }
            }
        },
        "internal_entry.Metadata": {
            "type": "object",
            "properties": {
                "fingerprint": {
                    "type": "string"
                },
                "storageBackend": {
                    "$ref": "#/definitions/nbox_internal_prefix.StorageBackendType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "nbox_internal_prefix.StorageBackendType": {
            "type": "string",
            "enum": [
                "dynamodb",
                "parameterstore",
                "parameterstore_secure"
            ],
            "x-enum-varnames": [
                "BackendDynamoDB",
                "BackendParameterStore",
                "BackendParameterStoreSecure"
            ]
        },
        "problem.Detail": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence.",
                    "type": "string"
                },
                "instance": {
                    "description": "Instance is a URI identifying the specific occurrence of the problem.",
                    "type": "string"
                },
                "requestId": {
                    "description": "RequestID correlates the problem to a specific request.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code for this occurrence.",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "Timestamp records when the problem was created.",
                    "type": "string"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.\nShould not change across occurrences of the same problem.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI identifying the problem type. Should provide\ndeveloper-readable documentation.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Bearer token authentication. Enter your JWT token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}